Rename to has/presets/dojo.ts to has/preset.ts
