+++ tests/widget-core/support/sendEvent.ts
@@ -1,4 +1,4 @@
-import has from '../../../src/has/presets/dojo';
+import has from '../../../src/has/preset';
 import { add } from '../../../src/has/has';
 import { deepAssign } from '../../../src/core/util';
 import global from '../../../src/shim/global';
